Change structure of packages, move mapper from main class to specific one. Change main class name
