+++ src/main/docker/Dockerfile.jvm
@@ -7,7 +7,7 @@
 #
 # Then, build the image with:
 #
-# docker build -f src/main/docker/Dockerfile.jvm -t quarkus/corpo-app-jvm .
+# docker build -f src/App/docker/Dockerfile.jvm -t quarkus/corpo-app-jvm .
 #
 # Then run the container using:
 #
