+++ src/main/docker/Dockerfile.native
@@ -7,7 +7,7 @@
 #
 # Then, build the image with:
 #
-# docker build -f src/main/docker/Dockerfile.native -t quarkus/corpo-app .
+# docker build -f src/App/docker/Dockerfile.native -t quarkus/corpo-app .
 #
 # Then run the container using:
 #
