+++ src/main/java/com/corpoapp/corpoevent/api/mapper/BaseMapper.java
@@ -1,0 +1,10 @@
+package com.corpoapp.corpoevent.api.mapper;
+
+import org.modelmapper.ModelMapper;
+
+import javax.enterprise.context.ApplicationScoped;
+
+@ApplicationScoped
+public class BaseMapper extends ModelMapper{
+
+}
