+++ src/main/java/com/corpoapp/corpoevent/domain/service/impl/CorporationServiceImpl.java
@@ -1,11 +1,11 @@
-package com.corpoapp.corpoevent.service.impl;
+package com.corpoapp.corpoevent.domain.service.impl;
 
-import com.corpoapp.corpoevent.dto.CorporationDTO;
-import com.corpoapp.corpoevent.entity.Corporation;
-import com.corpoapp.corpoevent.entity.User;
-import com.corpoapp.corpoevent.exceptions.mapper.UserException;
-import com.corpoapp.corpoevent.service.CorporationService;
-import com.corpoapp.corpoevent.service.UserService;
+import com.corpoapp.corpoevent.api.dto.CorporationDTO;
+import com.corpoapp.corpoevent.domain.entity.Corporation;
+import com.corpoapp.corpoevent.domain.entity.User;
+import com.corpoapp.corpoevent.api.mapper.exception.UserException;
+import com.corpoapp.corpoevent.domain.service.CorporationService;
+import com.corpoapp.corpoevent.domain.service.UserService;
 import org.modelmapper.ModelMapper;
 import org.modelmapper.TypeToken;
