+++ src/test/java/com/corpoapp/corpoevent/domain/service/UserServiceTest.java
@@ -1,7 +1,7 @@
-package com.corpoapp.corpoevent.service;
+package com.corpoapp.corpoevent.domain.service;
 
-import com.corpoapp.corpoevent.entity.User;
-import com.corpoapp.corpoevent.exceptions.mapper.UserException;
+import com.corpoapp.corpoevent.domain.entity.User;
+import com.corpoapp.corpoevent.api.mapper.exception.UserException;
 import io.quarkus.test.junit.QuarkusTest;
 import org.junit.Assert;
 import org.junit.jupiter.api.Assertions;
